fastAPI/{{ cookiecutter.project_name }}/requirements.txt
# basics
numpy
pandas
scikit-learn
pytest
cookiecutter
make
python-dotenv>=0.5.1
scipy==1.15
click
#statsmodels


# backend
fastapi[standard]
email-validator
passlib[bcrypt]
pyjwt
python-decouple
pymongo[srv]
pyarrow
uvicorn 
python-multipart
bcrypt==3.2.0



pendulum
loguru
groq
uv

pretty-errors